requirements.txt
# ICU Sepsis Dynamic Subtype Discovery via Self-Supervised Patient Trajectory Representation

# Dependencies

# --- Data Processing ---
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# --- Configuration ---
pyyaml>=6.0

# --- Machine Learning ---
scikit-learn>=1.3.0

# --- Dimensionality Reduction & Visualization ---
umap-learn>=0.5.3
matplotlib>=3.7.0

# --- Survival Analysis (optional, recommended) ---
lifelines>=0.27.0

# --- MIMIC-IV Data Pipeline ---
duckdb>=0.9.0

# --- Explainability (optional, for V2) ---
# shap>=0.42.0

# --- Deep Learning (for V2, not needed for MVP) ---
# torch>=2.0.0
# pytorch-lightning>=2.0.0

# --- Gradient Boosting (optional baseline) ---
# xgboost>=1.7.0
